batctl: Remove dead code from old sysfs parsing
The old batman-adv sysfs files provided all their available options
inside their output. This made it necessary to parse the output to
analyse it before we accept an input given to batctl.

This functionality was removed as it wasn't the correct style according
to the sysfs maintainer. Not all code related to that were removed right
away when batctl was adopted to the new sysfs handling.

Signed-off-by: Sven Eckelmann <sven.eckelmann@gmx.de>