batctl: Add an optional interval for watch-mode
If you are running a very mobile mesh network with a fast originator
interval, then the last-seen interval might not update fast enough
depending on the actual scenario, because the default polling interval
of batctl is 1 second. With this patch, an optional interval in seconds
and a precision of 3 digits before and after an optional dot can be
added, making the batctl-watch-mode updates faster.

Signed-off-by: Linus Lüssing <linus.luessing@web.de>