batman-adv: split tt_local/global_entry flags field for local and remote computations
The tt_global/local_entry 'flags' field contains flags used either for
local and remote computations. To optimise the usage of this field, its
length has been increased to 16bit in order to use the eight leading bits
(from 0 to 7) to store flags that have to be sent on the wire, while the
eight ending ones are used for local computation only.

Signed-off-by: Antonio Quartulli <ordex@autistici.org>