batctl: Human readable log levels instead of bitmasks
As the number of log levels increases, it becomes less user friendly to
assign a log level, as this is given as a bitmask in decimal form. This
patch changes this, so that one or more log levels can be given as
human readable words (e.g. "batctl ll batman routes").

Signed-off-by: Martin Hundebøll <martin@hundeboll.net>