batctl: Return failure code when ping was not successful
The standard iputils-ping returns an error code when it could not receive a
single reply to a echo request. batctl ping behaved differently and made it
only possible to detect that the target is not reachable by parsing the text
output of batctl.

A new return code EXIT_NOSUCCESS (2) is introduced to distinguish a general
failure from the inability to reach another node.

Reported-by: Filippo Sallemi <tonyputi@gmail.com>
Signed-off-by: Sven Eckelmann <sven@narfation.org>