batctl: Translate client mac to originator mac for ping/traceroute
A user may want to check the path selected by batman-adv for any client when
only the MAC of the device with the IPv4 is known, but not the originator
address. batctl can assist the user by parsing the global translation table
to automatically translate it for him.

The ping itself will only be done until the originator is reached and not until
the actual client could be contacted because the client cannot understand
batman-adv ping packets.

Signed-off-by: Sven Eckelmann <sven@narfation.org>
Signed-off-by: Marek Lindner <lindner_marek@yahoo.de>