batctl: Add tool to translate mac/IPv4/bat_host to originator address
Users may want to get the originator which is responsible for a specific mac
address. These addresses can be specified using IPv4 address, ethernet address
or bat-host name.

For example, the output can be used to get the TQ value from the originator
table when only the client address is known and not the gateway address.

Signed-off-by: Sven Eckelmann <sven@narfation.org>
Signed-off-by: Marek Lindner <lindner_marek@yahoo.de>