batctl: use ETH_P_BATMAN
ETH_P_BATMAN is now defined in linux/if_ether.h. batctl does not need
to use the private BATADV_ETH_P_BATMAN define.
However, depending on the kernel-headers version installed on the
system, this new constant may not be available, therefore a redefinition
inside an ifndef check is needed.

Signed-off-by: Antonio Quartulli <ordex@autistici.org>