batctl: man page - fix undefined macro warning
A line beginning with ' or . in a groff file has a special meaning and
cannot be used for literal text. Instead the interpreter will try to
execute it and fail doing so. A zero-width space or reformatting the
paragraph can be used as simple workaround.

This problem was detected using: LC_ALL=en_US.UTF-8 MANROFFSEQ=''
MANWIDTH=80 man --warnings -E UTF-8 -l -Tutf8 -Z man/batctl.8
>/dev/null

Reported-by: Pau Koning <paukoning@gmail.com>
Signed-off-by: Marek Lindner <lindner_marek@yahoo.de>