batctl: Fix warning about comparing signed vs. unsigned
The macros NLMSG_OK and NLMSG_NEXT operate on the parameter len as if it would
be a variable from the type size_t. Using ssize_t (signed size_t) can cause
warnings about comparing a signed with an unsigned variable.

It is ok to store the signed len variable in an unsigned variable because an
check is used to stop processing when it was negative.

Signed-off-by: Sven Eckelmann <sven@open-mesh.com>
Signed-off-by: Marek Lindner <mareklindner@neomailbox.ch>