batctl: Fix static buffer for resolved mac adddress
The resolve_mac_* functions are currently not reentrant-safe because they are
using ether_aton. Thus all functions returning a mac address in a similar way
must also ensure that their data is in a statically allocated buffer and not on
a stack.

resolve_mac_from_cache did that for the wrong buffer.

Signed-off-by: Sven Eckelmann <sven@narfation.org>
Signed-off-by: Marek Lindner <mareklindner@neomailbox.ch>