batctl: tcpdump - add OGM2 and ELP parsing support
These new packet types were introduced with B.A.T.M.A.N. V and
required tcpdump parsing support.
With this change it is now possible to properly read the content
these packets by means of the batctl tdpcump command.

In order to reduce code duplication, the mechanism used
to dump TVLV containers has been extracted and moved to a new
function.

Signed-off-by: Antonio Quartulli <antonio.quartulli@open-mesh.com>
Signed-off-by: Marek Lindner <mareklindner@neomailbox.ch>