batctl: Replace list implementation with linux-like-list
The list-batman implementation of linked lists has an interface which tries
to resample the linux double linked list implementation but in reality is
incompatible. This resulted in code which moves most of the pointer
arithmetic back to the actual code. This is rather unfortunate because this
abstraction should reduce the pointer arithmetic and not increase it.

Also the incompatibilities between the list implementation used in
batman-adv and the one used by batctl made coding features in both
components error prone because each component required a slightly different
implementation for the same list manipulation.

Signed-off-by: Sven Eckelmann <sven@narfation.org>