batctl: Use monotonic time source for icmp timing
gettimeofday is not monotonic. It would give wrong results in some
situation like leap seconds, when switching between daylight saving time
and non-DST, NTP changes the time or when the administrator of the system
changes it manually. The function is also obsolete since POSIX.1-2008.

clock_gettime is recommended to get a monotonic timesource and as general
replacement of gettimeofday.

Signed-off-by: Sven Eckelmann <sven@narfation.org>