batctl: optchar variable uses wrong type
The variable "optchar" used char instead of int leading to a non working
batctl tp command on some platforms. These platforms are using ABIs with
unsigned as default char type.

getopt will return a negative value whenever it is finished parsing. But
these platforms will then store it in an unsigned char and thus are not
able to interpret later as end of the parsing loop (-1). The tpmeter code
will therefore always print the usage information and terminate batctl with
an error code.

Signed-off-by: Andreas Pape <apape@phoenixcontact.com>
[sven@narfation.org: Replaced commit message]
Signed-off-by: Sven Eckelmann <sven@narfation.org>