batctl: Add command to create/destroy batman-adv interface
The command "create" can be used to create a batman-adv interface without
any interface attached. This is helpful when the interfaces should be
configured independently of the first attached interface.

The command "destroy" can be used to destroy a batman-adv interface without
going through all attached interfaces and delete them manually.

Signed-off-by: Sven Eckelmann <sven@narfation.org>
Signed-off-by: Simon Wunderlich <sw@simonwunderlich.de>