batctl: Add License-Filename to GPL-2.0 files
The FSFE REUSE practices [1] recommend to add a "License-Filename" tag to
files which don't contain the complete license text. The GPL-2.0 files
usually only have a small notice at the beginning of the file. The longer
license is usually stored in a separate file and therefore should be
referenced accordingly.

[1] https://reuse.software/practices/

Signed-off-by: Sven Eckelmann <sven@narfation.org>