batctl: Don't translate multicast and zero mac addresses
The IP to mac translation code already skips zero and multicast mac
addressses because there might be multiple entries in the global
translation table which would match this entry. There is no "best" entry
originator for such an entry and such the result is basically a random
pick.

Translating such an input mac address (from ping/traceoute/translate) to a
single originator address is therefore not appropriate.

Signed-off-by: Sven Eckelmann <sven@narfation.org>