batctl: Store usage line next to command
The single line usage information for a subcommand is the simplest way to
learn how to use a command. It is displayed when either the -h option is
used or when the initial part of the batctl command is incorrect.

Since the commands are no longer registered manually in the main function,
the usage line should also no longer be stored there. Instead add it to the
same file as the command itself.

Signed-off-by: Sven Eckelmann <sven@narfation.org>