batctl: Prepare command infrastructure for shared functions
The command structure first mapped from one name and abbrevation to a
single function of batctl. But there can be commands which share a
significant portion of the code. To support these kind of things better,
add an additional arg pointer to the command structure. The handler then
receive a state struct from the main function instead of the mesh_iface.

The command implementation can then use the state to retrieve both the
mesh_iface and the cmd with the arg pointer.

Signed-off-by: Sven Eckelmann <sven@narfation.org>