batctl: Simplify enabling for bisect_iv subcommand
The variable content of CONFIG_BATCTL_BISECT can be used to decide whether
the bisect_iv command is enabled or not. This would make it possible in the
future to introduce a generic way to enable or disable commands. Combining
this with link-time-optimization and/or function/data-section allows to
relatively freely scale the functionality of batctl and its size.

Signed-off-by: Sven Eckelmann <sven@narfation.org>