batctl: Add netlink fallback for sysfs' iface_status
The batman-adv kernel module can now be compiled without support for sysfs.
But the status of an interface in the list generated by `batctl interface`
can only get the status via the per hardif sysfs file iface_status. To
still have some information, fallback to BATADV_CMD_GET_HARDIF if the sysfs
file is not available.

Signed-off-by: Sven Eckelmann <sven@narfation.org>