batctl: Prefer netlink hardif status retrieval over sysfs
The sysfs code in batman-adv was changed to print a deprecated warning when
sysfs files are accessed. The `batctl if` call would therefore cause
warnings like this in the kernel log:

  batman_adv: [Deprecated]: batctl (pid 18540) Use of sysfs file "iface_status".
  Use batadv genl family instead

It is now appropriate to try the generic netlink BATADV_CMD_GET_HARDIF
request first to get the status of the interface before falling back to
sysfs.

Reported-by: Linus Lüssing <linus.luessing@c0d3.blue>
Signed-off-by: Sven Eckelmann <sven@narfation.org>