batctl: Replace '-m meshif' option with selector prefix
The '-m meshif' option doesn't follow the prefix style which is used for
VLANs/hard interfaces. It is also not clear for a user which command
understands the '-m' option and which is actually operating on a
global/non-meshif specific scope. To clean this up, an additional command
prefix called meshif is implemented:

  $ batctl meshif bat0 ...

This can be omitted when it would be "meshif bat0". And when it is
specified, all non-meshif specific subcommands will be rejected.

The old option '-m meshif' option is still available to avoid breaking
existing scripts. It will will show an error message but continues to work
for now. Users of batctl should still transition to the new format.

Signed-off-by: Sven Eckelmann <sven@narfation.org>