batctl: Allow to omit explicit prefix name
batctl allows three types of netdev based prefixes

* meshif <netdev>
* vlan <vdev>
* hardif <netdev>

The first word is used to tell batctl that the netdev is expected to be of
a specific type and all further information should be parsed in context of
this specific type. This avoids ambiguity when parsing the command line
information.

But there are various situations when there is no ambiguity at all. For all
of them, following points are true:

* <netdev|vdev> exists
* <netdev|vdev> has not the name of any existing batctl subcommand
* <netdev|vdev> is not "vlan", "vid", "hardif" or "meshif"

In these situations, the first word can be omitted and batctl can guess the
type automatically.

Signed-off-by: Sven Eckelmann <sven@narfation.org>