batctl: Disable automatic destruction of empty meshifs
The sysfs interface of batman-adv used an automatic drop of empty batadv
meshifs. This was necessary because no other method to remove this type
of interface was available. The same behavior was tried to emulate in the
batctl rtnetlink implementation to make the migration easier.

But this caused problems with both (rtnetlink/sysfs) backends because the
caller might not be aware of the current state of hardifs list for the
current meshif. The meshif might therefore be deleted accidentally (and its
configuration lost) even when the caller only wanted to swap some devices
around.

Instead print information about the necessary addition steps to drop the
empty meshif instead of performing it automatically.

Signed-off-by: Sven Eckelmann <sven@narfation.org>