batctl: Switch active routing algo list to netlink
The sysfs support is disabled by default in batman-adv since a while and
will be removed in 2021. The generic netlink interface should be used
instead. But the list of active routing algorithms was only available when
batman-adv was build with sysfs support.

Instead of walking through (not existing) sysfs entries, query the list of
batadv interfaces via rtnl and use the already existing helpers to get the
routing algorithm.

Signed-off-by: Sven Eckelmann <sven@narfation.org>