batctl: Realign netlink_policy array
The netlink_policy contains an array of structures with several members in
it. But the initialization of the entries of this array was not aligned
like any other initialization in the batctl code - instead it was tried to
have everything on the same same line. This resulted in rather long lines
and could result in larger commits which just try to realign the whole
table when a new entry with a long identifier was added.

Just realign the array to better match the rest of the code.

Signed-off-by: Sven Eckelmann <sven@narfation.org>