batctl: Avoid boolean in structures
Booleans inside structures to really only store a single bit of information
is often a waste of space because a boolean is stored using multiple bytes.
While this is not too relevant for batctl, just use the same way of storing
such values as in the rest of batman-adv to avoid different paradigms in
related code.

Signed-off-by: Sven Eckelmann <sven@narfation.org>