batctl: Introduce JSON_* command types
The JSON_MIF, JSON_HIF and JSON_VID subcommand types will provide access to the various
batadv generic netlink query types. This will allow a consumer to query the
state of a batadv interface (and its hard-interfaces/vlans) without the
need to:

* understand the details of generic netlink and the batadv family
* parse the text response of the debug tables or other subcommands

These commands can become a way to retrieve the information in a machine
readable way but they must be considered at the moment as "experimental"
and changes in the structure might still happpen.

Signed-off-by: Alexander Sarmanow <asarmanow@gmail.com>
[sven@narfation.org: fix commit message, cleanup changes, split up in
 vid/mif/hif types, fix manpage]
Signed-off-by: Sven Eckelmann <sven@narfation.org>