batctl: man: Use native list support
The lists were manually implemented by shifting the content to the right and
then adding a prefix to it. This turned out to be the source of various
errors which affected the whole document after the "list".

The .IP list macro can be used instead to simplify these kind of lists.

Signed-off-by: Sven Eckelmann <sven@narfation.org>