batctl: man: Rewrite SEE ALSO list
The dot command is not relevant anymore to batctl because the vis command
is no longer provided by batctl/batman-adv.

The ip and ip-link are the main tools to handle the kernel net_devices via
rtnetlink. They are therefore a must for anyone trying to use the batadv
interfaces.

The trace-cmd is the tool which has to be used to gather the logs from the
kernel when batctl enabled the various loglevels for an batadv interface.

The bridge utility is the main tool to configure bridges over the
batman-adv interfaces. It must also be used to read out the mdb state in
case the multicast handling of batman-adv has to be debugged.

Signed-off-by: Sven Eckelmann <sven@narfation.org>